cleanup perturb cli test
